+++ src/main/java/com/example/course/entities/Lesson.java
@@ -15,6 +15,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Entity
 @Table(name = "lessons")
 @Data
@@ -25,7 +27,7 @@ public class Lesson {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     // Relationship to Course
     @ManyToOne(fetch = FetchType.LAZY)
