+++ src/main/java/com/example/course/repositories/LessonRepository.java
@@ -5,11 +5,12 @@ import com.example.course.entities.Lesson;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface LessonRepository extends JpaRepository<Lesson, Long> {
+public interface LessonRepository extends JpaRepository<Lesson, Integer> {
 
     Optional<List<Lesson>> findByCourse(Course course);
 }
